Allow spawn-fcgi in the foreground
Been looking for a clean way to spawn fastcgi listeners in the foreground and I
believe Trey has figured it out. This patch should take care of it by keeping
spawn-fcgi from losing track of the copied listening processes and allow full
supervison.

taken from http://rubyists.com/articles/2005/05/03/spawn-fcgi-in-the-foreground