Initialize the library on the specified socket instead of stdin
Initialization of the library could not be done without creating a
request on the standard input.

This can be a problem if the program has disabled the standard input
(a linux daemon for example).

Then this commit enables to initialize the library on a specified
socket without using the standard input.